#30 get the safe file transmission size as long type

+++ ngrinder-controller/src/test/java/org/ngrinder/perftest/service/MockPerfTestRunnable.java
@@ -48,7 +48,7 @@ public class MockPerfTestRunnable extends PerfTestRunnable {
 	}
 
 	@Override
-	protected int getSafeTransmissionThreshold() {
+	protected long getSafeTransmissionThreshold() {
 		return 1;
 	}
